[refactor] hooksをlibの下に移動、secToMMSSとかを切り出し
