+++ client/src/components/page/space/memberBlock/MemberBlockContainer.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useMembersValue } from "../../../../hooks/useSyncMembers";
+import { useMembersValue } from "../../../../lib/hooks/useSyncMembers";
 import { MemberBlock } from "./MemberBlock";
 
 export const MemberBlockContainer: React.VFC = () => {
