+++ client/src/components/page/space/timerBlock/TimerBlock.tsx
@@ -1,28 +1,12 @@
 import React from "react";
-import { Box, Button, Progress, Spacer, Text, VStack } from "@chakra-ui/react";
+import { Box, Progress, Spacer, Text, VStack } from "@chakra-ui/react";
+import { secToHHMMSS, secToMMSS } from "../../../../lib/secondToString";
 
 type TimerBlockProps = {
   remainSec: number;
   fullSec: number;
   sectionTitle: string;
   adminController?: React.ReactChildren;
-};
-
-const secToMMSS = (sec: number) => {
-  const mm = Math.floor(Math.abs(sec) / 60);
-  const ss = Math.abs(sec) % 60;
-  return `${sec < 0 ? "-" : ""}${mm.toString().padStart(2, "0")}:${ss
-    .toString()
-    .padStart(2, "0")}`;
-};
-
-const secToHHMMSS = (sec: number) => {
-  const hh = Math.floor(Math.abs(sec) / 3600);
-  const mm = Math.floor(Math.abs(sec) / 60) % 60;
-  const ss = Math.abs(sec) % 60;
-  return `${sec < 0 ? "-" : ""}${hh.toString().padStart(2, "0")}:${mm
-    .toString()
-    .padStart(2, "0")}:${ss.toString().padStart(2, "0")}`;
 };
 
 export const TimerBlock: React.VFC<TimerBlockProps> = ({
