+++ client/src/components/page/space/timetableBlock/TimetableBlockContainer.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { TimetableBlock } from "./TimetableBlock";
-import { useTimetableCardsProps } from "../../../../hooks/useSyncTimetable";
+import { useTimetableCardsProps } from "../../../../lib/hooks/useSyncTimetable";
 
 export const TimetableBlockContainer: React.VFC = () => {
   const cards = useTimetableCardsProps();
