+++ client/src/lib/hooks/useSyncTimer.ts
