+++ client/src/lib/hooks/useUserInRoom.ts
